+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 ### 1.1.3
     - Now supporting PHP-Rocker 1.1.0
     - Added method removeFile(name, callback)
+
+### 1.1.4
+    - Now possible to declare authentication mechanism when calling Rocker.setUser();
+    - More callback arguments in functions Rocker.me() Rocker.version() and Rocker.operations()